requirements.txt
# streamlit
streamlit
streamlit_js_eval

# langchain
langchain
langchainhub
langchain_community
langchain_openai

# vector database
chromadb

# opneai
openai

# docs
pypdf
pdfminer
pdfminer.six
pytesseract

# huggingface
huggingface_hub

# python dev
python-dotenv
requests
bs4
unstructured
youtube-transcript-api
pytube
unstructured[all-docs]
